+++ CHANGELOG.md
@@ -33,6 +33,7 @@ You should also include the user name that made the change.
 - AVIF support @tamaina
 - Add Cloudflare Turnstile CAPTCHA support @CyberRex0
 - Introduce retention-rate aggregation @syuilo
+- Make possible to export favorited notes @syuilo
 - Server: signToActivityPubGet is set to true by default @syuilo
 - Server: improve syslog performance @syuilo
 - Server: improve note scoring for featured notes @CyberRex0